explicitly construct QUrl for new (local) skipList entries
getExistingDirectory returns a string, which we implicitly constructed
a QUrl from without using fromLocalFile() resulting in a QUrl that is
isValid && !isEmpty, but when converted toLocalFile() it will be "" as it
is !isLocal.

To resolve this problem simply construct a QUrl::fromLocalFile explicitly
to make sure that the resulting url is valid, !empty and local.

Otherwise this will attempt to insert empty strings as paths, rendering the
feature defunct.

REVIEW: 126985