use toLocalFile instead of path of the QUrl to fix Win32 compat
path() would be /foo irregardless of platform, on Windows that's /C:/ which
is not correct as the local representation is C:/. toLocalFile handles this
for us.

I've also changed the var name to be more representitive of what it
contains and is used for i.e. checking the localPath. should remote URL
validation be introduced at a later point it would then be obvious that one
should not be using that particular variable ;)