do not check paths on remotes (and crashing in doing so)
path="" == path="/"

and a path cannot not start with a slash anyway, so the check is moot.

"sftp://localhosthome/me/" != "sftp://localhost/home/me/"

it also does nothing to actually assert that the path is absolute as

"sftp://localhost/home/me/../you/" is relative but would pass the slash
check.

Also, we resolve relative remote paths just fine. Would be weird if we
didn't.