ditch dead code
kiconeffect was noop in kdelibs(4?) and eventually removed entirely. so
we haven't had it in use for 10 or so years. it was removed without
replacement, so we have nothing to port the disabled code to. Instead
drop it. given no one really complained about it in 10 years it's probably
not vital