do not confuse portable seperator with native ones
url.toLocalFile returns C:/ so we incorrectly turned that into C:/\\
i.e. C:// in portable speak. this then resulted in the filtering in
locallister not working because it's doing a startsWith check between
the ignore (e.g. C:/Windows) and the incoming path (e.g. C:/\\) that
would naturally fail.

so, to fix this we'll simply want to stop using qdir::separator

BUG: 450863