add a go to overview action
previously there was no real way to get back to the mainpage. it's all a
bit awkward because of how state management works, but I don't
realistically see that improving a lot with qwidgets. in kirigami we can
do it smarter while porting, if we ever port

in the meantime this adds an overview button to the toolbar which shows
the summary widget

BUG: 450512