grand qml rewrite
this moves the entire main ui (excluding settings) to qtquick from
qwidgets.

this loses some of the more esoteric kxmlgui features but should be a
good path forward on the balance

the actual map is still being qpainted, so this isn't super ideal but
the plan there is to hopefully use kquickcharts eventually instead of a
custom implementation.

settings dialog still needs porting but is kept qwidget for now to at
least reduce the scope of this port slightly

BUG: 443751
BUG: 445018
BUG: 445026
BUG: 445027
BUG: 449688