replace previous summary with tidy overview widget
it occurred to us that the old disk view was fishy on a number of levels

- it had a non exhaustive filter list of bad mounts
- this then raised the question why the mounts are missing (when e.g.
the user expected them to be there)
- the free space on a partition is not reflective of the space filelight
will be able to determine the use of (e.g. when +x is missing on a dir
that dir will be missing from the fileview)
- when talking about bind mounts (such as inside a sandbox) the two
sizes will almost never add up
- when inside a sandbox potentially only one valid path is available
($home) so having the previous view was slightly weird

the grand new overview widget simply displays three buttons to get right
into scanning root, home, or a custom path. root is hidden inside a
sandbox.

this preserves most of the spirit of what the previous view wanted to do
while scaling better and doing away with judgment calls on the edge
cases such as bind mounts

BUG: 452275
BUG: 440152
BUG: 440151
BUG: 426594