fix remote listing
this was crashing because we randomly casted a qobject to qwidget
without any safety nets whatsoever. I've refactored the ctor arguments
to not be nonesense and removed the offending widget setting (we have no
widget don't you know)

additionally the remote scanning was skipping running management in
scanmanager. it's not part of running() and accordingly emits
runningChanged when needed