refactor the remotelister to not rely on raw pointers so much
safer this way. there are various crashes on errors.ubuntu.com and the
microsoft partner center that ultimately boil down to the raw pointer
shuffling we are doing everywhere. so we should be getting away from
that since clearly we aren't doing a good job of managing them and the
modern cpp approach is to not use owning raw pointers anyway so the less
we do it the better

also propagate is now looping instead of descending in on itself, the
latter risks running into stack exhaustion for particularly deep
directory trees. to support that it's also become a free function so we
don't accidentally look at the wrong pointer (i.e. `this`)