Fix the path shown for unrepresented files
To represent the bunch of smaller files, we use a fake `File`
with a custom string as the filename.

Add a specific subclass for this case, where we also set the
parent `Folder` so the right path is shown. Otherwise,
`File::displayName()` will show the current working
directory as the path for our fake item.

BUG: 461964
FIXED-IN: 22.12.0