make qml cache disabling conditional on statvfs
this is still a bit risky but 5GiB hopefully gives us enough leeway to
create valid caches before potentially running out of space. e.g. when
the system is currently leaking onto a disk somehow, the space could not
survive for very long. the hope is that this is enough of an edge case
that it should not ever happen in practice.

filesystem::space should be supported on all supported platforms