posix: undo apfs hack
when I introduced this I thought it would have no adverse affects but it
indeed breaks sparse file handling where the st_blocks are really 0 and
not just fake 0.
undo the hack and consider fuse-apfs bugged. there certainly seems to be
no other file system that has blocks set to 0 all the time