tmp windows long path support
I am kinda giving up on this because its sensitivity to duplicated
backslashes is driving me nuts. we'd have to make sure that strings
coming into the walker are always with final separator or without. but
that is also complicated by the fact that the input may be forward or
backward slashes.... all in all terrible