manually position tooltip so it's correct
Qt tries to reposition when there isn't enough space and gloriously
fails. Instead adjust the coordinates such that Qt has no reason to
reposition.

since this adds a lot of code we now have a shared shapetooltip item
that is used by the centershape and the segmentshape. also correctly set
shapeitem as parent for our tooltip (the coordindates are relative to it
because our mouseyarea is on top of shapeitem)

BUG: 487339