Only show footer on map page
On the overview page, all it does is show "No files" and "Aborting scan"
which amount to useless visual noise. Refactor to no longer show any
status messages there, and then only show the footer on the map page
where it can show anything meaningful.

Also, port it to use QQC.ToolBar, which is the correct UI element to use
for this, instead of assembling something by hand.