make makeMap and closeURL more reliable
previously it failed to manipulate the stack correctly because of
various changes in the Qt6 version. overall the new code should be more
reliable now.

we now somewhat consistently pop back to the intialPage and push the
mappage on top. in case of new construction we also manually construct
the page instead of letting kirigami do the work so we have control over
the page's lifetime

BUG: 485744