windowswalker: walk right over errors
when an error occurs we must not process the entry because the fileinfo
points to undefined data. simply continue after logging the error.

this fixes a problem with corrupted folders that would produce entires
of themselves because the '.' and '..' skip logic didn't apply. so we
would produce (unexpected) 'foo/.' entry listings.
the locallister would then iterate into the subfolder '.' and that would
fail but yield an entry and the lister'd iterate into subfolder './.'
and that would fail ... kept going until we exhaust the path length