memory: prefer shared_ptr for qobject + deleteLater on them
introduce a new make_shared_qobject helper that creates a shared_ptr
with deleteLater semantics. this notably improves safety when used as
class members (that may be used in slot invocations) to prevent
use-after-free scenarios.

unique_ptr has the disadvantage that its deleter signature matters to
the type. it results in very noisy code and complicates assignment of
nullptr_ts. shared_ptr meanwhile does not care. as such it seems
preferrable.

this fixes a crash where we deleted the RemoteLister through resetting
it to nullptr in the updateRunning lambda. this would then delete the
kdirlister and its internals, leaving the signal call chain in an
inconsistent state and crashing when returning the call chain to
KCoreDirListerPrivate::CachedItemsJob::done (which was then working with
an already deleted this)

BUG: 507656