Don't update hovered item when contextMenu is visible
Let's keep the item highlighted to which the context menu refers.

It gives me each time a heart attack when I'm about to permanently
delete an item and the highlight of Filelight already changed to another
item that I definitely don't want to delete. When we stop updating the
highlight when the contextMenu is visible the highlight stays with
the items the context actions refer to.